Share the masonry layout and apply it to the curating index
The curating index still used the old single-<ul> `.creating-grid` markup,
which broke once that class became a flex column container. Extract the
responsive round-robin masonry (column count + left-to-right distribution)
into a reusable <MasonryGrid> and use it for both the creating and curating
indexes, so galleries get the same newest-across-the-top mosaic.
